Добавил Resource.kt для заворачивания ответов Проверил работспособность и содержимое запросов/ответов Чуть разбросал архитектуру, пора заходить в DI
