+++ app/src/main/java/ru/practicum/android/diploma/network/HeadHunterRepository.kt
@@ -1,32 +1,24 @@
 package ru.practicum.android.diploma.network
 
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import ru.practicum.android.diploma.network.api.HeadHunterApplicationApi
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flow
 import ru.practicum.android.diploma.network.api.Locale
+import ru.practicum.android.diploma.network.dto.HeadHunterRequest
+import ru.practicum.android.diploma.network.dto.responses.LocaleResponse
+import ru.practicum.android.diploma.utils.Resource
 
 const val BASE_URL = "https://api.hh.ru"
 
 class HeadHunterRepository {
+    private val client = RetrofitBasedClient()
 
-    private val retrofit =
-        Retrofit.Builder()
-            .baseUrl(BASE_URL)
-            .addConverterFactory(GsonConverterFactory.create())
-            .client(createOkHttpDefaultClient())
-            .build()
-    private val serverService = retrofit.create(HeadHunterApplicationApi::class.java)
-    suspend fun getLocales(): List<Locale> {
-        return serverService.getLocales()
+    suspend fun getLocales(): Flow<Resource<List<Locale>>> = flow {
+        val response = client.doRequest(HeadHunterRequest.LocalesList)
+        if (response.resultCode == 200) {
+            emit(Resource.Success((response as LocaleResponse).localeList))
+        } else {
+            emit(Resource.Error(""))
+        }
     }
 
-    private fun createOkHttpDefaultClient(): OkHttpClient {
-        val interceptor = HttpLoggingInterceptor()
-        interceptor.setLevel(HttpLoggingInterceptor.Level.BODY)
-        return OkHttpClient().newBuilder()
-            .addInterceptor(interceptor)
-            .build()
-    }
 }
