+++ app/src/main/java/ru/practicum/android/diploma/network/api/HeadHunterNetworkClient.kt
@@ -4,5 +4,5 @@ import ru.practicum.android.diploma.network.dto.responses.Response
 import ru.practicum.android.diploma.network.dto.HeadHunterRequest
 
 interface HeadHunterNetworkClient {
-    suspend fun doRequest(dto: HeadHunterRequest): Response
+    suspend fun doRequest(request: HeadHunterRequest): Response
 }
