+++ app/src/main/java/ru/practicum/android/diploma/utils/Resource.kt
@@ -1,0 +1,6 @@
+package ru.practicum.android.diploma.utils
+
+sealed class Resource<T>(val data: T? = null, val message: String? = null) {
+    class Success<T>(data: T) : Resource<T>(data)
+    class Error<T>(message: String) : Resource<T>(message = message)
+}
